move unit tests in their own file
