+++ crates/primitives/src/transaction/eip2930.rs
@@ -65,3 +65,63 @@ impl TxEip2930 {
         self.input.len() // input
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::TxEip2930;
+    use crate::{
+        transaction::{signature::Signature, TransactionKind},
+        Address, Bytes, Transaction, TransactionSigned, U256,
+    };
+    use bytes::BytesMut;
+    use reth_rlp::{Decodable, Encodable};
+
+    #[test]
+    fn test_decode_create() {
+        // tests that a contract creation tx encodes and decodes properly
+        let request = Transaction::Eip2930(TxEip2930 {
+            chain_id: 1u64,
+            nonce: 0,
+            gas_price: 1,
+            gas_limit: 2,
+            to: TransactionKind::Create,
+            value: 3,
+            input: Bytes::from(vec![1, 2]),
+            access_list: Default::default(),
+        });
+        let signature = Signature { odd_y_parity: true, r: U256::default(), s: U256::default() };
+        let tx = TransactionSigned::from_transaction_and_signature(request, signature);
+
+        let mut encoded = BytesMut::new();
+        tx.encode(&mut encoded);
+        assert_eq!(encoded.len(), tx.length());
+
+        let decoded = TransactionSigned::decode(&mut &*encoded).unwrap();
+        assert_eq!(decoded, tx);
+    }
+
+    #[test]
+    fn test_decode_call() {
+        let request = Transaction::Eip2930(TxEip2930 {
+            chain_id: 1u64,
+            nonce: 0,
+            gas_price: 1,
+            gas_limit: 2,
+            to: TransactionKind::Call(Address::default()),
+            value: 3,
+            input: Bytes::from(vec![1, 2]),
+            access_list: Default::default(),
+        });
+
+        let signature = Signature { odd_y_parity: true, r: U256::default(), s: U256::default() };
+
+        let tx = TransactionSigned::from_transaction_and_signature(request, signature);
+
+        let mut encoded = BytesMut::new();
+        tx.encode(&mut encoded);
+        assert_eq!(encoded.len(), tx.length());
+
+        let decoded = TransactionSigned::decode(&mut &*encoded).unwrap();
+        assert_eq!(decoded, tx);
+    }
+}
